Time: 381 ms (5%), Space: 167.6 MB (18.45%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool judgeSquareSum(int c) {
-        unordered_map<int, int> ump;
+        unordered_map<long, int> ump;
 
         for(int i=0; i<=sqrt(c); i++){
             ump[i*i]++;
